Arrumado problema com ponto flutuante

+++ app/src/main/java/com/example/tcc/model/ChildData.kt
@@ -3,9 +3,9 @@ package com.example.tcc.model
 data class ChildData(
     var category: String?,
     var description: String?,
-    var quantity: Long?,
+    var quantity: Double?,
     var unity: String?,
-    var price: Long?,
+    var price: Double?,
     var type: String?,
-    var total: Long?
+    var total: Double?
 )
