+++ app/src/main/java/com/example/tcc/model/Property.kt
@@ -1,5 +1,5 @@
 package com.example.tcc.model
 
 class Property(var name: String?,
-               var dimension: Long?,
+               var dimension: Double?,
                var location: String?)
